moved vis, align and dynamic logic to Text component
